Shooter for 2 players
This is a 2D python shooter using the tkinter library. The object of the game is to hit your opponent and try to avoid him hitting you.
The game is designed for two players. One plays on the left side of the keyboard, the other on the right (arrows or using NumLock)
All the rules are in the game itself.
Good game